+++ Sources/Development/app.swift
@@ -1,11 +1,20 @@
 import Vapor
 
-public func app(_ environment: Environment) throws -> Application {
-    let app = Application(environment: environment) {
-        var s = Services.default()
-        try configure(&s)
-        return s
+final class Development: ApplicationDelegate {
+    init() { }
+    
+    func configure(_ s: inout Services) {
+        s.extend(Routes.self) { r, c in
+            try routes(r, c)
+        }
+
+        s.register(HTTPServer.Configuration.self) { c in
+            switch c.environment {
+            case .tls:
+                return .init(hostname: "127.0.0.1", port: 8443, tlsConfiguration: tls)
+            default:
+                return .init(hostname: "127.0.0.1", port: 8080)
+            }
+        }
     }
-    try boot(app)
-    return app
 }
